chore: delete dead save/ package, move AnthropicSession to v2/llm
- knes-agent-tools/save/: delete KnesSave, MinimapTracker, SaveFormatCodec
  and their tests — zero usages anywhere
- AnthropicSession: knes.agent.llm → knes.agent.v2.llm (alongside
  AnthropicHttp, HaikuClient, SonnetClient); closes orphaned llm package
- EmulatorToolsetStepModeTest: LocalEmulatorToolset(session) after interface split
